Newuser

+++ src/main/java/com/sayone/ebazzar/service/ProductService.java
@@ -4,6 +4,8 @@ import com.sayone.ebazzar.document.ElasticProduct;
 import com.sayone.ebazzar.dto.ProductDto;
 import com.sayone.ebazzar.entity.ProductEntity;
 import com.sayone.ebazzar.entity.SubCategoryEntity;
+import com.sayone.ebazzar.exception.ErrorMessages;
+import com.sayone.ebazzar.exception.RequestException;
 import com.sayone.ebazzar.repository.ProductRepository;
 import com.sayone.ebazzar.repository.SubCategoryRepository;
 import org.springframework.beans.BeanUtils;
@@ -43,7 +45,8 @@ public class ProductService {
     }
 
 
-    public ProductDto addProduct(ProductDto addProduct) {
+    public ProductDto addProduct(ProductDto addProduct,String userType) {
+        if(!userType.equals("seller")) throw new RequestException(ErrorMessages.INVALID_SELLER.getErrorMessages());
         ProductEntity product = new ProductEntity();
         ProductDto returnValue = new ProductDto();
 
@@ -82,7 +85,8 @@ public class ProductService {
     }
 
 
-    public ProductDto updateProduct(ProductDto body, Long Id) {
+    public ProductDto updateProduct(ProductDto body, Long Id,String userType) {
+        if(!userType.equals("seller")) throw new RequestException(ErrorMessages.INVALID_SELLER.getErrorMessages());
         ProductEntity product = new ProductEntity();
         BeanUtils.copyProperties(body, product);
         ProductEntity productToUpdate = productRepository.findByProductId(Id);
@@ -100,7 +104,8 @@ public class ProductService {
         return returnValue;
     }
 
-    public void deleteProduct(Long id) {
+    public void deleteProduct(Long id,String userType) {
+        if(!userType.equals("seller")) throw new RequestException(ErrorMessages.INVALID_SELLER.getErrorMessages());
         productRepository.deleteById(id);
     }
 }
